+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Simon Petrus Changelog
 
+## v0.0.3 (3) --- 2024-08-09
+
+- New: Added loading window during uploads and most other operations
+- Improved: The Qt main window no longer freezes during operations
+
 ## v0.0.2 (2) --- 2024-08-09
 
 - New: Added "Warta Jemaat" and "Tata Ibadah" automatic uploader